requirements.txt
openai==0.28
discord.py==2.3.2
